requirements.txt
asgiref==3.5.2; python_version >= "3.8"
django==4.0.4; python_version >= "3.8"
djangorestframework-simplejwt==5.2.0; python_version >= "3.7"
djangorestframework==3.13.1; python_version >= "3.6"
pyjwt==2.4.0; python_version >= "3.7"
pytz==2022.1; python_version >= "3.7"
sqlparse==0.4.2; python_version >= "3.8"
tzdata==2022.1; sys_platform == "win32" and python_version >= "3.8"
